# backend/openapi.yaml
components:
  schemas:
...
    Category:
      properties:
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        created_at:
          format: date-time
          title: Created At
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon
        id:
          title: Id
          type: integer
        name:
          title: Name
          type: string
      required:
      - name
      - id
      - created_at
      title: Category
      type: object
    CategoryCreate:
      properties:
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon
        name:
          title: Name
          type: string
      required:
      - name
      title: CategoryCreate
      type: object
    CategoryUpdate:
      properties:
        color:
          anyOf:
          - type: string
          - type: 'null'
          title: Color
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      title: CategoryUpdate
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    Token:
      properties:
        access_token:
          title: Access Token
          type: string
        token_type:
          default: bearer
          title: Token Type
          type: string
      required:
      - access_token
      title: Token
      type: object
    UserCreate:
      properties:
        aadhar_card:
          anyOf:
          - type: string
          - type: 'null'
          title: Aadhar Card
        account_details:
          anyOf:
          - type: string
          - type: 'null'
          title: Account Details
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        age:
          anyOf:
          - type: integer
          - type: 'null'
          title: Age
        company_details:
          anyOf:
          - type: string
          - type: 'null'
          title: Company Details
        company_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Company Id
        dob:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Dob
        educational_qualification:
          anyOf:
          - type: string
          - type: 'null'
          title: Educational Qualification
        email:
          format: email
          title: Email
          type: string
        emergency_contact:
          anyOf:
          - type: string
          - type: 'null'
          title: Emergency Contact
        full_name:
          title: Full Name
          type: string
        gender:
          anyOf:
          - type: string
          - type: 'null'
          title: Gender
        is_teacher:
          default: false
          title: Is Teacher
          type: boolean
        marital_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Marital Status
        parents_contact_details:
          anyOf:
          - type: string
          - type: 'null'
          title: Parents Contact Details
        parents_email:
          anyOf:
          - type: string
          - type: 'null'
          title: Parents Email
        password:
          title: Password
          type: string
        phone_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone Number
        preferred_language:
          anyOf:
          - type: string
          - type: 'null'
          title: Preferred Language
        role:
          title: Role
          type: string
        seller_record:
          anyOf:
          - type: string
          - type: 'null'
          title: Seller Record
        seller_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Seller Type
        sub_role:
          title: Sub Role
          type: string
        year_of_experience:
          anyOf:
          - type: integer
          - type: 'null'
          title: Year Of Experience
      required:
      - email
      - password
      - full_name
      - role
      - sub_role
      title: UserCreate
      type: object
    UserLogin:
      properties:
        email:
          format: email
          title: Email
          type: string
        password:
          title: Password
          type: string
      required:
      - email
      - password
      title: UserLogin
      type: object
    UserOut:
      properties:
        aadhar_card:
          anyOf:
          - type: string
          - type: 'null'
          title: Aadhar Card
        account_details:
          anyOf:
          - type: string
          - type: 'null'
          title: Account Details
        address:
          anyOf:
          - type: string
          - type: 'null'
          title: Address
        age:
          anyOf:
          - type: integer
          - type: 'null'
          title: Age
        company_details:
          anyOf:
          - type: string
          - type: 'null'
          title: Company Details
        company_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Company Id
        dob:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Dob
        educational_qualification:
          anyOf:
          - type: string
          - type: 'null'
          title: Educational Qualification
        email:
          format: email
          title: Email
          type: string
        emergency_contact:
          anyOf:
          - type: string
          - type: 'null'
          title: Emergency Contact
        full_name:
          title: Full Name
          type: string
        gender:
          anyOf:
          - type: string
          - type: 'null'
          title: Gender
        gyan_coins:
          title: Gyan Coins
          type: integer
        id:
          title: Id
          type: integer
        is_teacher:
          title: Is Teacher
          type: boolean
        marital_status:
          anyOf:
          - type: string
          - type: 'null'
          title: Marital Status
        parents_contact_details:
          anyOf:
          - type: string
          - type: 'null'
          title: Parents Contact Details
        parents_email:
          anyOf:
          - type: string
          - type: 'null'
          title: Parents Email
        phone_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone Number
        preferred_language:
          anyOf:
          - type: string
          - type: 'null'
          title: Preferred Language
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
        seller_record:
          anyOf:
          - type: string
          - type: 'null'
          title: Seller Record
        seller_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Seller Type
        sub_role:
          anyOf:
          - type: string
          - type: 'null'
          title: Sub Role
        year_of_experience:
          anyOf:
          - type: integer
          - type: 'null'
          title: Year Of Experience
      required:
      - id
      - email
      - full_name
      - age
      - gender
      - role
      - sub_role
      - educational_qualification
      - preferred_language
      - phone_number
      - address
      - emergency_contact
      - aadhar_card
      - account_details
      - dob
      - marital_status
      - year_of_experience
      - parents_contact_details
      - parents_email
      - seller_type
      - company_id
      - seller_record
      - company_details
      - is_teacher
      - gyan_coins
      title: UserOut
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: /api/auth/login
      type: oauth2
info:
  title: Gyanvruksh API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/admin/users:
    get:
      description: Get all users with optional filtering
      operationId: get_all_users_api_admin_users_get
      parameters:
      - in: query
        name: skip
        required: false
        schema:
          default: 0
          title: Skip
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          title: Limit
          type: integer
      - in: query
        name: role
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Get All Users
      tags:
      - admin
...
  /api/admin/users/{user_id}:
    delete:
      description: Delete a user (soft delete by deactivating)
      operationId: delete_user_api_admin_users__user_id__delete
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Delete User
      tags:
      - admin
...
  /api/auth/admin/create-admin:
    post:
      operationId: create_admin_api_auth_admin_create_admin_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Create Admin
      tags:
      - auth
  /api/auth/admin/users:
    get:
      operationId: list_users_api_auth_admin_users_get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserOut'
                title: Response List Users Api Auth Admin Users Get
                type: array
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: List Users
      tags:
      - auth
  /api/auth/admin/users/{user_id}:
    delete:
      operationId: delete_user_api_auth_admin_users__user_id__delete
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Delete User
      tags:
      - auth
    put:
      operationId: update_user_api_auth_admin_users__user_id__put
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Update User
      tags:
      - auth
  /api/auth/login:
    post:
      operationId: login_api_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Login
      tags:
      - auth
  /api/auth/logout:
    post:
      operationId: logout_api_auth_logout_post
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Logout
      tags:
      - auth
  /api/auth/me:
    get:
      operationId: me_api_auth_me_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Me
      tags:
      - auth
    put:
      description: Update current user's profile
      operationId: update_profile_api_auth_me_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Update Profile
      tags:
      - auth
  /api/auth/register:
    post:
      operationId: register_api_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Register
      tags:
      - auth
  /api/categories/:
    get:
      description: Get all categories
      operationId: get_categories_api_categories__get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Category'
                title: Response Get Categories Api Categories  Get
                type: array
          description: Successful Response
      summary: Get Categories
      tags:
      - categories
    post:
      description: Create a new category (Admin only)
      operationId: create_category_api_categories__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Create Category
      tags:
      - categories
  /api/categories/{category_id}:
    delete:
      description: Delete a category (Admin only)
      operationId: delete_category_api_categories__category_id__delete
      parameters:
      - in: path
        name: category_id
        required: true
        schema:
          title: Category Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Delete Category
      tags:
      - categories
    get:
      description: Get a specific category by ID
      operationId: get_category_api_categories__category_id__get
      parameters:
      - in: path
        name: category_id
        required: true
        schema:
          title: Category Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Category
      tags:
      - categories
    put:
      description: Update a category (Admin only)
      operationId: update_category_api_categories__category_id__put
      parameters:
      - in: path
        name: category_id
        required: true
        schema:
          title: Category Id
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Update Category
      tags:
      - categories
...